# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HealthPass Data Submission API",
    "description": "Swagger definition for HealthPass data submission API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/datasubmission",
      "description": "CDT"
    }
  ],
  "paths": {
...
    "/organization": {
      "put": {
        "tags": [
          "organization"
        ],
        "summary": "Update organization",
        "description": "Update data for registered entity",
        "operationId": "updateOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Registered entity data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEntityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully updated registered entity data"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Get a list of all registered organizations",
        "description": "Get a list of all registered organizations",
        "operationId": "getOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved registered organizations"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/organization/{regEntity}/displayschemaid": {
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Get organization's display schema id",
        "description": "Get organization's display schema id",
        "operationId": "getDisplaySchemaID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved display schema id"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/organization/{regEntity}/consentReceipt/{holderId}": {
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Get organization's consent receipt for the holder",
        "description": "Get organization's consent receipt for the holder",
        "operationId": "getConsentReceipt",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "holderId",
            "in": "path",
            "description": "Holder id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved consent receipt"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/organization/{regEntity}/consentRevoke/{holderId}": {
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Get organization's consent revoke for the holder",
        "description": "Get organization's consent revoke for the holder",
        "operationId": "getConsentRevoke",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "holderId",
            "in": "path",
            "description": "Holder id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved consent revoke"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/organization/{regEntity}": {
      "delete": {
        "tags": [
          "organization"
        ],
        "summary": "Delete organization",
        "description": "Delete data for registered entity",
        "operationId": "deleteOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted registered entity data"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RegisterEntityRequest": {
        "type": "object",
        "example": {
          "entity": "organization_name",
          "entityType": "holder-upload",
          "name": "organization_name",
          "contact": "support@org.com",
          "website": "https://www.org.com/",
          "privacyPolicy": "https://www.org.com/about-us/privacy-policy",
          "userAgreement": "TBD",
          "services": [
            {
                "service": "upload",
                "purposes": [
                    {
                        "purpose": "TBD"
                    }
                ]
            }
          ],
          "consentInfo": {
            "version": "ISO/IEC 29184:2020",
            "jurisdiction": "US",
            "language": "en",
            "collectionMethod": "direct",
            "consentType": {
              "explicit": true
            },
            "termination": "Instructions for how to terminate",
            "thirdPartyTransfer": false,
            "principal": {
              "id": "",
              "key": ""
            },
            "controller": {
              "name": "organization_name",
              "address": {
                "line": "12345 Rockville rd",
                "city": "NYC",
                "state": "NY",
                "postalCode": "12345",
                "country": "United States"
              },
              "email": "support@org.com",
              "website": "https://www.org.com"
            },
            "privacyPolicy": "https://www.org.com/privacypolicy",
            "services": [
              {
                "description": "ORG service",
                "purposes": [
                  {
                    "description": "ORG purpose TBD.",
                    "purposeCategory": {
                      "research": true,
                      "service": true
                    },
                    "piiCategory": {
                      "medical": true
                    }
                  }
                ]
              }
            ],
            "consentId": "9de1154e-b110-4072-9fa1-adab960b150c",
            "consentTimestamp": 1620227004
          },
          "profileSchema": "did:hpass:59cd606341eb4a4a6c1a25d94a5f842ecf83ccd441dbda8abcd9274c9acd9334:67cba75b1719b5efba1addd32602f827fd378f2654288b1a4e381f8dddf40af3;id=holder-upload-contact;version=0.1",
          "userSchema": "did:hpass:59cd606341eb4a4a6c1a25d94a5f842ecf83ccd441dbda8abcd9274c9acd9334:67cba75b1719b5efba1addd32602f827fd378f2654288b1a4e381f8dddf40af3;id=ncl-upload-id;version=0.3",
          "displaySchema": "did:hpass:59cd606341eb4a4a6c1a25d94a5f842ecf83ccd441dbda8abcd9274c9acd9334:67cba75b1719b5efba1addd32602f827fd378f2654288b1a4e381f8dddf40af3;id=ncl-upload-id;version=0.3",
          "profileCredentialExpiry": 31536000,
          "userCredentialExpiry": 31536000,
          "issuerId": "hpass.issuer1",
          "mapper": {
            "profileSchema": {
              "name": "name",
              "contact": "contact",
              "website": "website",
              "services": "services",
              "privacyPolicy": "privacyPolicy",
              "userAgreement": "userAgreement"
            },
            "userSchema": {
              "name": {
                "familyName": "familyName",
                "givenName": "givenName"
              },
              "location": "location",
              "passengerNumber": "123456",
              "id": "publicKey",
              "registrationCode": "registrationCode"
            },
            "consentSchema": {
              "version": "version",
              "jurisdiction": "jurisdiction",
              "language": "language",
              "collectionMethod": "collectionMethod",
              "principal": "principal",
              "controller": "controller",
              "privacyPolicy": "privacyPolicy",
              "consentType": "consentType",
              "termination": "termination",
              "thirdPartyTransfer": "thirdPartyTransfer",
              "services": "services"
            },
            "consentRevokeSchema": {
              "version": "version",
              "jurisdiction": "jurisdiction",
              "language": "language",
              "principal": "principal",
              "controller": "controller"
            }
          },
          "userRegistrationConfig": {
            "org": "organization_name",
            "entityType": "holder-upload",
            "flow": {
              "registrationCodeAuth": true,
              "mfaAuth": false,
              "showUserAgreement": true,
              "showRegistrationForm": true
            },
            "userAgreement": "<b>user agreement</b> and <i>Terms and Conditions</i> text goes here.",
            "registrationForm": {
              "givenName": {
                "type": "string",
                "displayType": "Text",
                "order": "02",
                "required": true,
                "minLength": 2,
                "maxLength": 32,
                "visible": true
              },
              "id": {
                "type": "string",
                "displayType": "Text",
                "visible": false
              },
              "registrationCode": {
                "type": "string",
                "displayType": "Text",
                "visible": false
              },
              "familyName": {
                "type": "string",
                "displayType": "Text",
                "order": "03",
                "required": true,
                "minLength": 2,
                "maxLength": 32,
                "visible": true
              },
              "passengerNumber": {
                "type": "string",
                "displayType": "Text",
                "order": "01",
                "required": true,
                "minLength": 2,
                "maxLength": 16,
                "visible": true
              },
              "location": {
                "type": "string",
                "displayType": "Text",
                "order": "04",
                "pattern": "^[a-zA-Z0]{2}[\\-][a-zA-Z0-9]+$",
                "required": true,
                "minLength": 2,
                "maxLength": 2,
                "visible": true
              }
            }
          }
        }
      },
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "InvalidRequestError": {
        "description": "Missing or invalid required fields in request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "InvalidGetRequestError": {
        "description": "Invalid fields in request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
    }
  }
}